requirements.txt
torch>=2.5.0
numpy>=2.0.0
scipy>=1.13.0
scikit-learn>=1.6.0
pandas>=2.3.0
tqdm>=4.67.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-cov>=5.0.0
